Isolated the scraping logic
Isolated the scraping logic so it can be used in tests
